feat: Namespace the new FF code under `PostHog.FeatureFlags`
This will make it more ergonomic for future updates. Most of the code stayed the same.

I've also added a helper `checked!` function that avoids the whole `{:ok, _}` check, but it's more prone to errors. People should still use `check` most of the time, but `check!` is more ergonomic for scripts
